server/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
requests==2.26.0
requests-oauthlib==1.3.0
